refactor: re-introduce `Session`
